Remove UserController.java
