--- src/main/java/com/group3/backend/controller/UserController.java
@@ -1,112 +1,0 @@
-package com.group3.backend.controller;
-
-import com.group3.backend.dto.request.LoginRequest;
-import com.group3.backend.dto.request.RegistrationRequest;
-import com.group3.backend.dto.response.AuthResponse;
-import com.group3.backend.model.User;
-import com.group3.backend.service.JwtService;
-import com.group3.backend.service.UserService;
-import jakarta.validation.Valid;
-import jakarta.servlet.http.Cookie;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.BadCredentialsException;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.web.bind.annotation.*;
-import org.springframework.http.ResponseEntity;
-
-@RestController
-@RequestMapping("/auth")
-@RequiredArgsConstructor
-public class UserController {
-
-    private final UserService service;
-
-    private final JwtService jwtService;
-
-    private final AuthenticationManager authenticationManager;
-
-    @GetMapping("/welcome")
-    public String welcome() {
-        return "Welcome this endpoint is not secure";
-    }
-
-    @PostMapping("/signup")
-    public ResponseEntity<String> signup(@Valid @RequestBody RegistrationRequest registrationRequest) {
-        try {
-            String result = service.registerUser(registrationRequest);
-            return ResponseEntity.status(HttpStatus.CREATED).body(result);
-        } catch (IllegalArgumentException e) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Registration Error: " + e.getMessage());
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("An unexpected error occurred during user registration: " + e.getMessage());
-        }
-    }
-
-    @PostMapping("/signin")
-    public ResponseEntity<?> signin(@RequestBody LoginRequest authRequest, HttpServletResponse response) {
-        Authentication authentication;
-        try {
-            authentication = authenticationManager.authenticate(
-                new UsernamePasswordAuthenticationToken(authRequest.getEmail(), authRequest.getPassword())
-            );
-        } catch (UsernameNotFoundException | BadCredentialsException e) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Invalid email or password.");
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-                    .body("An unexpected error occurred: " + e.getMessage());
-        }
-
-        if (authentication.isAuthenticated()) {
-            User user = service.getUserByEmail(authRequest.getEmail());
-            String accessToken = jwtService.generateAccessToken(user.getEmail(), user.getId(), user.getRole().getName().name());
-            String refreshToken = jwtService.generateRefreshToken(user.getEmail());
-
-            // Set refresh token as HttpOnly cookie
-            Cookie cookie = new Cookie("refreshToken", refreshToken);
-            cookie.setHttpOnly(true);
-            cookie.setPath("/");
-            cookie.setMaxAge(7 * 24 * 60 * 60); // 7 days
-            
-            response.addCookie(cookie);
-
-            return ResponseEntity.ok(new AuthResponse(accessToken, refreshToken, user.getEmail(), user.getRole().getName().name(), user.getId()));
-        } else {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Authentication failed!");
-        }
-    }
-
-    @PostMapping("/refresh")
-    public ResponseEntity<?> refreshToken(@CookieValue(value = "refreshToken", required = false) String refreshToken) {
-        if (refreshToken == null || refreshToken.isEmpty()) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Refresh token is missing");
-        }
-
-        try {
-            String email = jwtService.extractEmail(refreshToken);
-            User user = service.getUserByEmail(email);
-
-            if (!jwtService.isTokenExpired(refreshToken)) {
-                String newAccessToken = jwtService.generateAccessToken(user.getEmail(), user.getId(), user.getRole().getName().name());
-
-                // Trả về access token mới, giữ nguyên refresh token
-                return ResponseEntity.ok(
-                    new AuthResponse(newAccessToken, refreshToken, email, user.getRole().getName().name(), user.getId())
-                );
-            } else {
-                return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Refresh token expired");
-            }
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Invalid refresh token");
-        }
-    }
-
-    @GetMapping("/validate")
-    public ResponseEntity<String> validateToken() {
-        return ResponseEntity.ok("Token is valid");
-    }
-}
